Add library struct to application, move util functions to impl
