+++ src/ui/artist_list.rs
@@ -1,7 +1,7 @@
 use crate::{
     application::Application,
     config,
-    library_utils::{artists_from_library, play_artist},
+    library_utils::play_artist,
     models::ArtistModel,
     ui::{
         artist::Artist,
@@ -89,8 +89,7 @@ impl ArtistList {
     }
 
     pub fn pull_artists(&self) {
-        let library = self.get_application().library().clone();
-        let artists = artists_from_library(&library.borrow());
+        let artists = self.get_application().library().artists_from_library();
         if artists.is_empty() {
             self.set_empty(true);
         } else {
